+++ src/main/java/com/ydays/dart/gamePlayer/GamePlayer.java
@@ -12,7 +12,7 @@ public class GamePlayer{
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @ManyToOne
     @JoinColumn(name="player_id", nullable=false)
@@ -32,7 +32,7 @@ public class GamePlayer{
 
     public GamePlayer(){}
 
-    public GamePlayer(int id, Player player, Game game, int remainingShots, int score, int order, int rank) {
+    public GamePlayer(Long id, Player player, Game game, int remainingShots, int score, int order, int rank) {
         this.id = id;
         this.player = player;
         this.game = game;
@@ -42,11 +42,12 @@ public class GamePlayer{
         this.rank = rank;
     }
 
-    public int getId() {
+
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
