+++ src/main/java/com/ydays/dart/joueur/Player.java
@@ -12,7 +12,7 @@ public class Player {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     private String name;
 
@@ -20,7 +20,7 @@ public class Player {
 
     public Player(){}
 
-    public Player(int id, String name, String email) {
+    public Player(Long id, String name, String email) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -29,11 +29,11 @@ public class Player {
     @OneToMany(mappedBy="player")
     private Set<GamePlayer> gamePlayers;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
